requirements.txt
# Bibliotecas Core
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
scipy>=1.9.0

# Visualização
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Machine Learning Avançado
xgboost>=1.6.0
lightgbm>=3.3.0
catboost>=1.1.0
imbalanced-learn>=0.9.0

# Interpretabilidade
shap>=0.41.0
lime>=0.2.0

# Feature Engineering
category-encoders>=2.5.0
feature-engine>=1.4.0

# Otimização de Hiperparâmetros
optuna>=3.0.0
hyperopt>=0.2.7

# Utilitários
joblib>=1.1.0
tqdm>=4.64.0
pyyaml>=6.0

# Jupyter e Desenvolvimento
jupyter>=1.0.0
ipywidgets>=7.7.0

# Validação e Testes
pytest>=7.0.0
pytest-cov>=3.0.0

# API
fastapi>=0.110.0
uvicorn>=0.25.0
